fix(ken-478): Address review - skills-only render, capture.rs [no-changelog]
#1. position()'s rendered join is now the ItemKind::Skill arm alone, which
is what the issue's Done-when names. Every other kind keeps dir.join(name)
byte-for-byte, so the diff provably changes nothing for a command reaching
that arm through exits::for_row. A new test pins both halves.

#2. Ratchet: no raise this round, and adopt.rs comes down rather than up.

read_tree and its budget move to engine/capture.rs. That is a real seam,
not a cut to make a number fit: it takes a path and returns bytes, its
three callers (adopt's plain and shared captures, fork's edited capture,
the package diff) read it as a black box, and it imports nothing from
adopt. Two of those callers were already reaching through
engine::adopt::read_tree for something that was never about adoption. Its
own test moves with it. Bodies move unchanged; the diff is the module
line, the use lines, and the paths at three call sites.

That leaves adopt.rs at 683, below the 735 it stood at before this issue,
so its row falls 763 -> 683. This is not the seam KEN-549 (e9bde7ea)
rejected: that one was adopt_shared.rs, refused because pointing its
imports one way meant moving read_tree and local_item_path into a file
about shared folders, where neither belongs. Lifting read_tree clear of
both is the move that was actually available.

error.rs stays at 423 and the typed AdoptNameUnusable stays with it. One
enum is the whole of a crate's error vocabulary, so it has no seam to
split at, and its row is raised for exactly this in the two most recent
commits that touched it: 402 -> 410 (3053d28d) and 410 -> 416 (4fef6f8e,
"error.rs ratchet baseline rises by its one new variant").
