+++ crates/core/src/engine/mod.rs
@@ -9,6 +9,7 @@ use crate::model::Scope;
 
 pub mod adopt;
 pub(crate) mod bundles;
+pub(crate) mod capture;
 mod catalog;
 mod config_edits;
 mod copilot;
